feat: hide and show password
